LayoutTests/fast/js/parse-error-external-script-in-new-Function.html

<html>
<head>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}
</script>
<script src="resources/parse-error-external-script-in-new-Function.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.notifyDone();
}
</script>
</head>
<body>
This tests that a line number and error message is printed for a parse error in a separate source file using new Function().
</body>
</html>
