LayoutTests/platform/chromium/fast/js/parse-error-external-script-in-new-Function-expected.txt

CONSOLE MESSAGE: line 4: Uncaught SyntaxError: Unexpected identifier
This tests that a line number and error message is printed for a parse error in a separate source file using new Function().
